move things out of scrape --> networkdriver
