+++ docs/scrapli/driver/core/cisco_nxos/driver.html
@@ -26,7 +26,7 @@
 <span>Expand source code</span>
 </summary>
 <pre><code class="python">&#34;&#34;&#34;scrapli.driver.core.cisco_nxos.driver&#34;&#34;&#34;
-from typing import Any, Dict
+from typing import Any, Callable, Dict, Union
 
 from scrapli.driver import NetworkDriver
 from scrapli.driver.network_driver import PrivilegeLevel
@@ -99,12 +99,22 @@ PRIVS = {
 
 
 class NXOSDriver(NetworkDriver):
-    def __init__(self, auth_secondary: str = &#34;&#34;, **kwargs: Dict[str, Any]):
+    def __init__(
+        self,
+        auth_secondary: str = &#34;&#34;,
+        session_pre_login_handler: Union[str, Callable[..., Any]] = &#34;&#34;,
+        session_disable_paging: Union[str, Callable[..., Any]] = &#34;terminal length 0&#34;,
+        **kwargs: Dict[str, Any],
+    ):
         &#34;&#34;&#34;
         NXOSDriver Object
 
         Args:
             auth_secondary: password to use for secondary authentication (enable)
+            session_pre_login_handler: callable or string that resolves to an importable function to
+                handle pre-login (pre disable paging) operations
+            session_disable_paging: callable, string that resolves to an importable function, or
+                string to send to device to disable paging
             **kwargs: keyword args to pass to inherited class(es)
 
         Returns:
@@ -113,7 +123,9 @@ class NXOSDriver(NetworkDriver):
         Raises:
             N/A
         &#34;&#34;&#34;
-        super().__init__(auth_secondary, **kwargs)
+        super().__init__(
+            auth_secondary, session_pre_login_handler, session_disable_paging, **kwargs
+        )
         self.privs = PRIVS
         self.default_desired_priv = &#34;privilege_exec&#34;
         self.textfsm_platform = &#34;cisco_nxos&#34;
@@ -131,7 +143,7 @@ class NXOSDriver(NetworkDriver):
 <dl>
 <dt id="scrapli.driver.core.cisco_nxos.driver.NXOSDriver"><code class="flex name class">
 <span>class <span class="ident">NXOSDriver</span></span>
-<span>(</span><span>auth_secondary='', **kwargs)</span>
+<span>(</span><span>auth_secondary='', session_pre_login_handler='', session_disable_paging='terminal length 0', **kwargs)</span>
 </code></dt>
 <dd>
 <section class="desc"><p>NXOSDriver Object</p>
@@ -139,6 +151,12 @@ class NXOSDriver(NetworkDriver):
 <dl>
 <dt><strong><code>auth_secondary</code></strong></dt>
 <dd>password to use for secondary authentication (enable)</dd>
+<dt><strong><code>session_pre_login_handler</code></strong></dt>
+<dd>callable or string that resolves to an importable function to
+handle pre-login (pre disable paging) operations</dd>
+<dt><strong><code>session_disable_paging</code></strong></dt>
+<dd>callable, string that resolves to an importable function, or
+string to send to device to disable paging</dd>
 <dt><strong><code>**kwargs</code></strong></dt>
 <dd>keyword args to pass to inherited class(es)</dd>
 </dl>
@@ -158,12 +176,22 @@ class NXOSDriver(NetworkDriver):
 <span>Expand source code</span>
 </summary>
 <pre><code class="python">class NXOSDriver(NetworkDriver):
-    def __init__(self, auth_secondary: str = &#34;&#34;, **kwargs: Dict[str, Any]):
+    def __init__(
+        self,
+        auth_secondary: str = &#34;&#34;,
+        session_pre_login_handler: Union[str, Callable[..., Any]] = &#34;&#34;,
+        session_disable_paging: Union[str, Callable[..., Any]] = &#34;terminal length 0&#34;,
+        **kwargs: Dict[str, Any],
+    ):
         &#34;&#34;&#34;
         NXOSDriver Object
 
         Args:
             auth_secondary: password to use for secondary authentication (enable)
+            session_pre_login_handler: callable or string that resolves to an importable function to
+                handle pre-login (pre disable paging) operations
+            session_disable_paging: callable, string that resolves to an importable function, or
+                string to send to device to disable paging
             **kwargs: keyword args to pass to inherited class(es)
 
         Returns:
@@ -172,7 +200,9 @@ class NXOSDriver(NetworkDriver):
         Raises:
             N/A
         &#34;&#34;&#34;
-        super().__init__(auth_secondary, **kwargs)
+        super().__init__(
+            auth_secondary, session_pre_login_handler, session_disable_paging, **kwargs
+        )
         self.privs = PRIVS
         self.default_desired_priv = &#34;privilege_exec&#34;
         self.textfsm_platform = &#34;cisco_nxos&#34;
