+++ scrapli/driver/driver.py
@@ -6,12 +6,7 @@ from types import TracebackType
 from typing import Any, Callable, Dict, Optional, Tuple, Type, Union
 
 from scrapli.channel import CHANNEL_ARGS, Channel
-from scrapli.helper import (
-    get_external_function,
-    resolve_ssh_config,
-    resolve_ssh_known_hosts,
-    validate_external_function,
-)
+from scrapli.helper import resolve_ssh_config, resolve_ssh_known_hosts
 from scrapli.transport import (
     MIKO_TRANSPORT_ARGS,
     SSH2_TRANSPORT_ARGS,
@@ -55,8 +50,6 @@ class Scrape:
         comms_prompt_pattern: str = r"^[a-z0-9.\-@()/:]{1,32}[#>$]$",
         comms_return_char: str = "\n",
         comms_ansi: bool = False,
-        session_pre_login_handler: Union[str, Callable[..., Any]] = "",
-        session_disable_paging: Union[str, Callable[..., Any]] = "terminal length 0",
         ssh_config_file: Union[str, bool] = False,
         ssh_known_hosts_file: str = "",
         transport: str = "system",
@@ -86,10 +79,6 @@ class Scrape:
                 case insensitive is just a convenience factor so i can be lazy.
             comms_return_char: character to use to send returns to host
             comms_ansi: True/False strip comms_ansi characters from output
-            session_pre_login_handler: callable or string that resolves to an importable function to
-                handle pre-login (pre disable paging) operations
-            session_disable_paging: callable, string that resolves to an importable function, or
-                string to send to device to disable paging
             ssh_config_file: string to path for ssh config file, True to use default ssh config file
                 or False to ignore default ssh config file
             ssh_known_hosts_file: string to path for ssh known hosts file, True to use default known
@@ -133,10 +122,6 @@ class Scrape:
         self.comms_return_char: str = ""
         self.comms_ansi: bool = False
         self._setup_comms(comms_prompt_pattern, comms_return_char, comms_ansi)
-
-        self.session_pre_login_handler: Optional[Callable[..., Any]] = None
-        self.session_disable_paging: Union[str, Callable[..., Any]] = ""
-        self._setup_session(session_pre_login_handler, session_disable_paging)
 
         if transport not in ("ssh2", "paramiko", "system", "telnet"):
             raise ValueError(
@@ -285,103 +270,6 @@ class Scrape:
         if not isinstance(comms_ansi, bool):
             raise TypeError(f"comms_ansi should be bool, got {type(comms_ansi)}")
         self.comms_ansi = comms_ansi
-
-    def _setup_session(
-        self,
-        session_pre_login_handler: Union[str, Callable[..., Any]],
-        session_disable_paging: Union[str, Callable[..., Any]],
-    ) -> None:
-        """
-        Parse and setup session attributes
-
-        Args:
-            session_pre_login_handler: pre login handler to parse/set
-            session_disable_paging: disable paging to parse/set
-
-        Returns:
-            N/A  # noqa: DAR202
-
-        Raises:
-            TypeError: if invalid type args provided
-
-        """
-        if session_pre_login_handler:
-            self.session_pre_login_handler = self._set_session_pre_login_handler(
-                session_pre_login_handler
-            )
-        else:
-            self.session_pre_login_handler = None
-        if callable(session_disable_paging):
-            self.session_disable_paging = session_disable_paging
-        if not isinstance(session_disable_paging, str) and not callable(session_disable_paging):
-            raise TypeError(
-                "session_disable_paging should be str or callable, got "
-                f"{type(session_disable_paging)}"
-            )
-        if session_disable_paging != "terminal length 0":
-            try:
-                self.session_disable_paging = self._set_session_disable_paging(
-                    session_disable_paging
-                )
-            except TypeError:
-                self.session_disable_paging = session_disable_paging
-        else:
-            self.session_disable_paging = "terminal length 0"
-
-    @staticmethod
-    def _set_session_pre_login_handler(
-        session_pre_login_handler: Union[str, Callable[..., Any]]
-    ) -> Optional[Callable[..., Any]]:
-        """
-        Return session_pre_login_handler argument
-
-        Args:
-            session_pre_login_handler: callable function, or string representing a path to
-                a callable
-
-        Returns:
-            session_pre_login_handler: callable or default empty string value
-
-        Raises:
-            TypeError: if provided string does not result in a callable
-
-        """
-        if callable(session_pre_login_handler):
-            return session_pre_login_handler
-        if not validate_external_function(session_pre_login_handler):
-            LOG.critical(f"Invalid comms_pre_login_handler: {session_pre_login_handler}")
-            raise TypeError(
-                f"{session_pre_login_handler} is an invalid session_pre_login_handler function "
-                "or path to a function."
-            )
-        return get_external_function(session_pre_login_handler)
-
-    @staticmethod
-    def _set_session_disable_paging(
-        session_disable_paging: Union[Callable[..., Any], str]
-    ) -> Union[Callable[..., Any], str]:
-        """
-        Return session_disable_paging argument
-
-        Args:
-            session_disable_paging: callable function, string representing a path to
-                a callable, or a string to send to device to disable paging
-
-        Returns:
-            session_disable_paging: callable or string to use to disable paging
-
-        Raises:
-            TypeError: if provided string does not result in a callable
-
-        """
-        if callable(session_disable_paging):
-            return session_disable_paging
-        if not validate_external_function(session_disable_paging):
-            raise TypeError(
-                f"{session_disable_paging} is an invalid session_disable_paging function or path "
-                "to a function. Assuming this is string to send to disable paging."
-            )
-        return get_external_function(session_disable_paging)
 
     def _setup_ssh_args(
         self, ssh_config_file: Union[str, bool], ssh_known_hosts_file: Union[str, bool]
