+++ tests/functional/conftest.py
@@ -18,8 +18,6 @@ def base_driver():
         timeout_ops=5,
         comms_prompt_pattern=r"^[a-z0-9.\-@()/:]{1,32}[#>$]$",
         comms_ansi=False,
-        session_pre_login_handler="",
-        session_disable_paging="terminal length 0",
         transport="system",
     ):
         conn = Scrape(
@@ -34,8 +32,6 @@ def base_driver():
             timeout_ops=timeout_ops,
             comms_prompt_pattern=comms_prompt_pattern,
             comms_ansi=comms_ansi,
-            session_pre_login_handler=session_pre_login_handler,
-            session_disable_paging=session_disable_paging,
             transport=transport,
         )
         conn.open()
